// src/api/generated/model/aPIEntitiesMember.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GitLab REST API
 * GitLab REST API used to interact with a GitLab installation.
 * OpenAPI spec version: v4
 */
import type { APIEntitiesCustomAttribute } from './aPIEntitiesCustomAttribute';
import type { APIEntitiesUserBasic } from './aPIEntitiesUserBasic';
import type { APIEntitiesIdentity } from './aPIEntitiesIdentity';
import type { APIEntitiesScimIdentity } from './aPIEntitiesScimIdentity';
import type { APIEntitiesMemberRole } from './aPIEntitiesMemberRole';

export interface APIEntitiesMember {
  id?: number;
  username?: string;
  public_email?: string;
  name?: string;
  state?: string;
  locked?: boolean;
  avatar_url?: string;
  avatar_path?: string;
  custom_attributes?: APIEntitiesCustomAttribute[];
  web_url?: string;
  access_level?: string;
  created_at?: string;
  created_by?: APIEntitiesUserBasic;
  expires_at?: string;
  group_saml_identity?: APIEntitiesIdentity;
  group_scim_identity?: APIEntitiesScimIdentity;
  email?: string;
  is_using_seat?: string;
  override?: string;
  membership_state?: string;
  member_role?: APIEntitiesMemberRole;
}